Updated Concurrency Model for Population Initialization
Previously, a new grouting was launched for each new individual
generated within a population. This incurred significant overhead
despite limiting the number of active coroutines at any one time to the
number of cpu processing units on the current machine. The new method
generates a fixed number of go routines, but keeps them alive, and
continuously producing individuals, until the target population size is
reached. Control of this process is achieved through a buffered channel
indicating the pool of remaining individuals that still need to be
generated.